Moving event log search to App.js so state is built on app load, saved to redux store for app route consumption
